platform-mc: Widen terminus instance type to uint16_t
EntityInstance is uint16_t in the PDR types, but the terminus
instance field and accessors were uint8_t. Align them to avoid
implicit narrowing when comparing with entity instance values.

Made-with: Cursor

+++ platform-mc/terminus.hpp
@@ -345,13 +345,13 @@ class Terminus
     }
 
     /** @brief The setter to set terminus's instance */
-    void setInstance(uint8_t inst)
+    void setInstance(uint16_t inst)
     {
         instance = inst;
     }
 
     /** @brief The getter to get terminus's instance */
-    std::optional<uint8_t> getInstance()
+    std::optional<uint16_t> getInstance()
     {
         return instance;
     }
@@ -400,7 +400,7 @@ class Terminus
     std::string terminusName;
 
     /** @brief Terminus instance from configuration */
-    std::optional<uint8_t> instance;
+    std::optional<uint16_t> instance;
 
     std::vector<std::shared_ptr<SensorAuxiliaryNames>>
         sensorAuxiliaryNamesTbl{};
